+++ types/src/module/system/damage/types.d.ts
@@ -64,10 +64,11 @@ interface DamageIRBypassData {
     immunity: {
         ignore: ImmunityType[];
         downgrade: DowngradedImmunity[];
+        redirect: ImmunityRedirect[];
     };
     resistance: {
         ignore: IgnoredResistance[];
-        redirect: RedirectedResistance[];
+        redirect: ResistanceRedirect[];
     };
 }
 interface DowngradedImmunity {
@@ -80,9 +81,14 @@ interface IgnoredResistance {
     max: number;
 }
 /** A damage type to check against instead if the target would resist the actual damage type */
-interface RedirectedResistance {
-    from: ResistanceType;
-    to: ResistanceType;
+interface ImmunityRedirect {
+    from: Exclude<DamageType, "untyped">;
+    to: Exclude<DamageType, "untyped">;
+}
+/** A damage type to check against instead if the target would resist the actual damage type */
+interface ResistanceRedirect {
+    from: Exclude<DamageType, "untyped">;
+    to: Exclude<DamageType, "untyped">;
 }
 interface ResolvedDamageFormulaData extends DamageFormulaData {
     roll?: never;
@@ -127,4 +133,4 @@ interface SpellDamageTemplate extends BaseDamageTemplate {
 type AfflictionDamageTemplate = SpellDamageTemplate;
 type SimpleDamageTemplate = SpellDamageTemplate;
 type DamageTemplate = WeaponDamageTemplate | SpellDamageTemplate | AfflictionDamageTemplate | SimpleDamageTemplate;
-export type { AfflictionDamageTemplate, BaseDamageData, CriticalInclusion, DamageCategory, DamageCategoryRenderData, DamageCategoryUnique, DamageDamageContext, DamageDiceFaces, DamageDieSize, DamageFormulaData, DamageIRBypassData, DamageKind, DamagePartialTerm, DamageRollRenderData, DamageTemplate, DamageType, DamageTypeRenderData, MaterialDamageEffect, RedirectedResistance, SimpleDamageTemplate, SpellDamageTemplate, WeaponBaseDamageData, WeaponDamageTemplate, };
+export type { AfflictionDamageTemplate, BaseDamageData, CriticalInclusion, DamageCategory, DamageCategoryRenderData, DamageCategoryUnique, DamageDamageContext, DamageDiceFaces, DamageDieSize, DamageFormulaData, DamageIRBypassData, DamageKind, DamagePartialTerm, DamageRollRenderData, DamageTemplate, DamageType, DamageTypeRenderData, ImmunityRedirect, MaterialDamageEffect, ResistanceRedirect, SimpleDamageTemplate, SpellDamageTemplate, WeaponBaseDamageData, WeaponDamageTemplate, };